+++ timer/sil_timer.h
@@ -4,7 +4,7 @@
 #endif
 
 #if !(_POSIX_C_SOURCE >= 199309L)
-#error NEED _POSIX_C_VERSION >= 199309L for struct timespec
+#error NEED _POSIX_C_SOURCE >= 199309L for struct timespec
 #endif
 
 #include <time.h>
@@ -15,4 +15,4 @@ struct sil_timer {
 
 void sil_timer_reset(struct sil_timer *timer);
 void sil_timer_elapsed_ms(struct sil_timer *timer, long *ms);
-void sil_timer_elapsed_s_ms(struct sil_timer *timer, int *s, long *ms);
+void sil_timer_elapsed_s_ms(struct sil_timer *timer, int *s, int *ms);
